LPS-117658 Move DDMWebConfiguration to API
